refactor: organize components according to their context
