+++ src/components/Base/Avatar/types.ts
