+++ src/router/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import { useAuth } from "~/contexts/auth/AuthContext";
-import Loading from "~/components/Loading";
+import Loading from "~/components/Base/Loading";
 
 import AppRouter from "./app.routes";
 import AuthRouter from "./auth.routes";
