+++ src/screens/CreateAppointment/styles.ts
@@ -1,7 +1,7 @@
 import styled from "styled-components/native";
 import { RectButton } from "react-native-gesture-handler";
 
-import Avatar from "~/components/Avatar";
+import Avatar from "~/components/Base/Avatar";
 
 import { ProviderListItemProps, ProviderListItemNameProps } from "./types";
