+++ src/screens/ForgotPassword/index.tsx
@@ -5,8 +5,8 @@ import { Form } from "@unform/mobile";
 import { useNavigation } from "@react-navigation/native";
 import { ValidationError } from "yup";
 
-import Button from "~/components/Button";
-import Input from "~/components/Input";
+import Button from "~/components/Base/Button";
+import Input from "~/components/Base/Input";
 import getValidationErrors from "~/utils/getValidationErrors";
 
 import AuthScreenLayout from "~/components/Layout/AuthScreenLayout";
